refactor : Separate the types directory to distinguish request data
